pipeline_requirements.txt
tensorflow
tensorflow-datasets==4.8.2
pycld3==0.22